move RequestCounter class to the statistics package
